+++ plugins/07_Legion/CHANGELOG.md
@@ -1,3 +1,6 @@
+# v119
+* Added Shadowquill rare to Suramar
+
 # v98
 
 * Localization updates: esES